requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
mlflow==2.8.0
pyarrow==13.0.0
matplotlib==3.7.2
seaborn==0.12.2
jupyter==1.0.0
notebook==7.0.6
pyyaml==6.0.1
python-dotenv==1.0.0
scipy==1.11.3